requirements.txt
einops
numpy
opencv-python
Pillow
PyYAML
scikit-learn
scipy
tabulate
tensorboardX
torchvision
torchvision
